+++ firmware/application/CMakeLists.txt
@@ -319,7 +319,6 @@ set(CPPSRC
 	apps/ui_touch_calibration.cpp
 	apps/ui_touchtunes.cpp
 	apps/ui_weatherstation.cpp
-    apps/ui_keeloqtx.cpp
 	protocols/aprs.cpp
 	protocols/ax25.cpp
 	protocols/dcs.cpp
