+++ firmware/application/external/keeloqtx/ui_keeloqtx.hpp
@@ -9,12 +9,12 @@
 #include "radio_state.hpp"
 #include "encoders.hpp"
 #include "string_format.hpp"
-#include "../keeloq_file.hpp"
-#include "../keeloq_keystore.hpp"
+#include "../../keeloq_file.hpp"
+#include "../../keeloq_keystore.hpp"
 
 #define KEELOQ_HEADER "101010101010101010101010000000000"
 
-namespace ui {
+namespace ui::external_app::ui_keeloqtx {
 
 const std::string keeloq_fragments[2] = { "110", "100" };
